% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{add_row_to_transfer_matrix}
\alias{add_row_to_transfer_matrix}
\title{Add an entry to the transferMatrix, for cand_out}
\usage{
add_row_to_transfer_matrix(
  votes,
  weights,
  running,
  stillIn,
  cand_out,
  prevVotes,
  votesForTransfer,
  verbose,
  transfer_matrix
)
}
\description{
Add an entry to the transferMatrix, for cand_out
}
