% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{avDebug}
\alias{avDebug}
\title{Perform a get request to an endpoint on the iexcloud server, will show the complete url
including security token being sent to the server for debug}
\usage{
avDebug(endpoint)
}
\arguments{
\item{endpoint}{a string which will form the variable are of the endpoint URL}
}
\value{
parsed response data, this will usually be a list of key:value pairs from parsed json object
}
\description{
Perform a get request to an endpoint on the iexcloud server, will show the complete url
including security token being sent to the server for debug
}
