% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{avGet}
\alias{avGet}
\title{Perform a get request to an endpoint on the alpha_vantage server}
\usage{
avGet(endpoint)
}
\arguments{
\item{endpoint}{a string which will form the variable are of the endpoint URL}
}
\value{
parsed response data, this will usually be a list of key:value pairs from parsed json object
}
\description{
Perform a get request to an endpoint on the alpha_vantage server
}
