% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{avRaw}
\alias{avRaw}
\title{Perform a get request to an endpoint on the alphaVantage server}
\usage{
avRaw(endpoint)
}
\arguments{
\item{endpoint}{a string which will form the variable are of the endpoint URL
#' @return raw response object}
}
\description{
Perform a get request to an endpoint on the alphaVantage server
}
