% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballot.R
\name{ballot}
\alias{ballot}
\title{Interface for running irv on ballot cards.}
\usage{
ballot(..., map = NULL)
}
\arguments{
\item{...}{Each entry in the ballot. Elements without votes cast can be
noted with 0, -1, NA, or any number below 1.}

\item{map}{optional character mapping giving names of the voting options}
}
\value{
ballot object
}
\description{
Interface for running irv on ballot cards.
}
