% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borda.R
\name{borda}
\alias{borda}
\title{Borda count}
\usage{
borda(
  ballots,
  nseats = 1,
  variant = "standard",
  report = FALSE,
  report_path = ifelse(report, "borda_single_report.html", NULL)
)
}
\arguments{
\item{ballots}{The list of ballots.}

\item{nseats}{The number of seats available.}

\item{variant}{The variant of Borda to use. Must be one of '\code{standard}', '\code{modified}', or '\code{dowdall}'.
See \url{https://en.wikipedia.org/wiki/Borda_count}.}

\item{report}{Generate an HTML report of the vote.}

\item{report_path}{When \code{report} is TRUE, the filepath to which to save the generated report.}
}
\description{
Count votes using the Borda method. This uses the simple \code{nseats > 1} version rather than 
Quota Borda Count (using 1st pref quotas), because that doesn't generalise to \code{nseats = 1}.
}
