% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condorcet.R
\name{condorcet}
\alias{condorcet}
\title{Condorcet count}
\usage{
condorcet(
  ballots,
  report = FALSE,
  report_path = ifelse(report, "condorcet_single_report.html", NULL)
)
}
\arguments{
\item{ballots}{The list of ballots.}

\item{report}{Generate an HTML report of the vote.}

\item{report_path}{When \code{report} is TRUE, the filepath to which to save the generated report.}
}
\description{
Count votes using the Condorcet method: for \code{nseats = 1} only.
}
