% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_count_animation.R
\name{create_stv_count_gif}
\alias{create_stv_count_gif}
\title{STV count animated gif}
\usage{
create_stv_count_gif(
  count_result,
  out_gif_path,
  candidate_colour_dict = NULL,
  title = "STV count",
  frame_delay_ms = 10,
  plot_width = 3,
  plot_height = 2.5
)
}
\arguments{
\item{count_result}{The count_result output from function stv, or any list containing count_table, candidates, and quota.}

\item{out_gif_path}{Filepath to which to save the animated gif.}

\item{candidate_colour_dict}{An optional named character vector of candidates and colour values to pass to `ggplot`.}

\item{frame_delay_ms}{The gap between frames passed to the convert function, in milliseconds.}

\item{plot_width}{The output gif width, in inches.}

\item{plot_height}{The output gif height, in inches.}
}
\description{
Plot an animation of an STV count and save a gif
}
