% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_stv}
\alias{ensemble_stv}
\title{Run stv multiple times and return elected frequencies,
to handle ties being broken randomly.}
\usage{
ensemble_stv(votes, nseats, nensemble, report = FALSE,
  report_path = ifelse(report, "stv_ens_report.html", NULL))
}
\arguments{
\item{votes}{List of ballot objects, such as returned by read_votes_from_csv}

\item{nseats}{The number of seats available}

\item{nensemble}{The size of the ensemble to run}

\item{report}{If TRUE, generate an HTML report of the vote, 
showing the possible count pathways and overall result (default: FALSE)}

\item{report_path}{If report is TRUE, the file name of the report
(default: stv_ens_report.html)}
}
\value{
An STV_ens object, containing:
  \describe{
    \item{ensemble_result:}{a table of all candidates elected at least once
      in the ensemble, and the percentage of iterations in which they were elected}
    \item{pathways:}{a list of the unique count pathways generated by the ensemble}
}
}
\description{
Run stv multiple times and return elected frequencies,
to handle ties being broken randomly.
}
