% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{get_above_quota}
\alias{get_above_quota}
\title{Return the names of the candidates who are over the current quota}
\usage{
get_above_quota(fps, quota, weights)
}
\description{
Return the names of the candidates who are over the current quota
}

