% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_count_animation.R
\name{get_anim_positions_from_count_table}
\alias{get_anim_positions_from_count_table}
\title{Take the stv count table with numeric columns and return a tweenr
animated data frame to produce the dot plots for the count}
\usage{
get_anim_positions_from_count_table(num_count_table, candidates, quota)
}
\arguments{
\item{num_count_table}{The count table output from function stv.}

\item{candidates}{The list of candidates; also output from function stv.}
}
\value{
A list containing
  \describe{
    \item{animation_data:}{tweenr animated data frame}
    \item{votes:}{a ballots data frame of dimension nballots rows by num_rounds columns}
  }
}
\description{
Take the stv count table with numeric columns and return a tweenr
animated data frame to produce the dot plots for the count
}
