% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{get_current_table}
\alias{get_current_table}
\title{For verbose output, print the current votes leaderboard}
\usage{
get_current_table(stillIn, fps, weights)
}
\description{
For verbose output, print the current votes leaderboard
}
