% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{get_number_to_eliminate}
\alias{get_number_to_eliminate}
\title{Return the number of bottom candidates that can be dropped in this round}
\usage{
get_number_to_eliminate(ordered_current_votes, quota)
}
\description{
Return the number of bottom candidates that can be dropped in this round
}
