% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{get_stv_loser}
\alias{get_stv_loser}
\title{Return the lowest-voted candidate by FPs, resolving ties randomly}
\usage{
get_stv_loser(fps, weights)
}
\description{
Return the lowest-voted candidate by FPs, resolving ties randomly
}

