% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irv.R
\name{irv}
\alias{irv}
\title{Instant runoff voting.}
\usage{
irv(votes)
}
\arguments{
\item{votes}{A names list of order-of-preference vote vectors.}
}
\value{
An IRV object, containing:

   winner: the winning entry

   ???: more to come!
}
\description{
Instant runoff voting.
}
\examples{
votes <- list(
  dex = c("Ice Skating", "Unihoc", "Food"),
  dean = c("Ice Skating", "Unihoc", "Food"),
  paul = c("Whiskey Tasting", "Established"),
  james = c("Ice Skating", "Unihoc", "Food")
)

irv_soln <- irv(votes)
irv_soln$winner
}

