% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{print_current_table}
\alias{print_current_table}
\title{For verbose output, print the current votes table}
\usage{
print_current_table(stillIn, fps, weights)
}
\description{
For verbose output, print the current votes table
}
