% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{read_votes_from_csv}
\alias{read_votes_from_csv}
\title{Read in a csv file of candidates and ballots in standard format, 
with candidates as rows and ballots as columns, and no header,
or with candidates as columns and ballots as rows}
\usage{
read_votes_from_csv(file_name, ballots_as_rows = FALSE)
}
\arguments{
\item{file_name}{The csv file containing votes}
}
\value{
List of ballots, named ballot1, ballot2, etc.
}
\description{
Read in a csv file of candidates and ballots in standard format, 
with candidates as rows and ballots as columns, and no header,
or with candidates as columns and ballots as rows
}
