% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schulze.R
\name{schulze}
\alias{schulze}
\title{Schulze count}
\usage{
schulze(
  ballots,
  nseats = 1,
  report = FALSE,
  report_path = ifelse(report, "schulze_single_report.html", NULL)
)
}
\arguments{
\item{ballots}{The list of ballots.}

\item{nseats}{The number of seats available.}

\item{report}{Generate an HTML report of the vote.}

\item{report_path}{When \code{report} is TRUE, the filepath to which to save the generated report.}
}
\description{
Count votes using the Schulze method. For \code{nseats = 1} this is identical to \code{condorcet}.
}
