% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{stv}
\alias{stv}
\title{Single transferable vote}
\usage{
stv(votes, nseats, verbose = FALSE, use_fps_for_final_tie = TRUE,
  getMatrix = FALSE, report = FALSE, report_path = ifelse(report,
  "stv_single_report.html", NULL), getTable = ifelse(report, TRUE, FALSE))
}
\arguments{
\item{votes}{A list of order-of-preference vote vectors, or a list of ballot
objects.}

\item{nseats}{How many seats to fill.}

\item{verbose}{Print out intermediate information or not}
}
\value{
An STV object, containing:
  \describe{
    \item{winners:}{the winning entries in order of preference}
    \item{More to come!}{}
  }
}
\description{
Single transferable vote
}
\examples{
votes <- list(
  dex = c("Ice Skating", "Unihoc", "Food"),
  dean = c("Ice Skating", "Unihoc", "Food"),
  paul = c("Whiskey Tasting", "Established"),
  james = c("Ice Skating", "Unihoc", "Food")
)

stv(votes, 2)

map <- c("e", "f", "i", "u", "w")
votes <- list(
  ballot(0, 3, 1, 2, 0, map = map),
  ballot(0, 3, 1, 2, 0, map = map),
  ballot(2, 0, 0, 0, 1, map = map),
  ballot(0, 3, 1, 2, 0, map = map)
)
stv(votes, 2)
}
