% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series_daily_adjusted}
\alias{time_series_daily_adjusted}
\title{returns market end of day adjusted data for a specified symbol}
\usage{
time_series_daily_adjusted(symbol, outputsize = "compact", lastN = 0)
}
\arguments{
\item{symbol}{string, a market data symbol such as "AMZN"}

\item{outputsize}{"full"; full returns ~ 20 years worth of data; compact returns last 100 data points}

\item{lastN}{0; if >0 lastN species how many of the most recent trading days will be returned}

\item{interval}{"1min"; choices are  1min, 5min, 15min, 30min, 60min}
}
\value{
dataframe
}
\description{
returns market end of day adjusted data for a specified symbol
}
