% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series_intraday}
\alias{time_series_intraday}
\title{returns intraday market data for a specified symbol}
\usage{
time_series_intraday(symbol, outputsize = "full", interval = "1min",
  lastday = FALSE)
}
\arguments{
\item{symbol}{string, a market data symbol such as "AMZN"}

\item{outputsize}{"full"; full returns ~ 5 days worth of data; compact returns last 100 data points}

\item{interval}{"1min"; choices are  1min, 5min, 15min, 30min, 60min}

\item{lastday}{FALSE; if TRUE only most recent trading days days will be returned}
}
\value{
dataframe
}
\description{
returns intraday market data for a specified symbol
}
