% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irv.R
\name{irv}
\alias{irv}
\title{Instant runoff voting.}
\usage{
irv(votes)
}
\arguments{
\item{votes}{A list of order-of-preference vote vectors, or a list of ballot
objects.}
}
\value{
An IRV object, containing:

   winner: the winning entry or entries in the case of a tie

   nrounds: the number of rounds required to find a winner

   rem_rounds: list of which entries were in contention at each round

   fps_rounds: table of first preference votes through each round

   eliminations: list of eliminated entries at each round
}
\description{
Instant runoff voting.
}
\examples{
votes <- list(
  dex = c("Ice Skating", "Unihoc", "Food"),
  dean = c("Ice Skating", "Unihoc", "Food"),
  paul = c("Whiskey Tasting", "Established"),
  james = c("Ice Skating", "Unihoc", "Food")
)

irv(votes)

map <- c("e", "f", "i", "u", "w")
votes <- list(
  ballot(0, 3, 1, 2, 0, map = map),
  ballot(0, 3, 1, 2, 0, map = map),
  ballot(2, 0, 0, 0, 1, map = map),
  ballot(0, 3, 1, 2, 0, map = map)
)
irv(votes)
}

