% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irv.R
\name{irv}
\alias{irv}
\title{Instant runoff voting.}
\usage{
irv(votelist)
}
\arguments{
\item{votelist}{A names list of order-of-preference vote vectors.}
}
\value{
An IRV object, containing:

   winner: the winning entry

   ???: more to come!
}
\description{
Instant runoff voting.
}

