% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irv.R
\name{irv}
\alias{irv}
\title{Instant runoff voting.}
\usage{
irv(votes, tiebreak = "all")
}
\arguments{
\item{votes}{A list of order-of-preference vote vectors, or a list of ballot
objects.}

\item{tiebreak}{String informing tiebreak behavior.
\describe{
  \item{"all"}{All tied losers will be dropped}
  \item{"random"}{Tied losers will be dropped randomly}
  \item{"nested"}{Tied losers will be broken by running a nested IRV only
                  considering votes for the losers.}
}}
}
\value{
An IRV object, containing:
  \describe{
    \item{winner:}{the winning entry or entries in the case of a tie}
    \item{nrounds:}{the number of rounds required to find a winner}
    \item{rem_rounds:}{list of which entries were in contention at each round}
    \item{fps_rounds:}{table of first preference votes through each round}
    \item{eliminations:}{list of eliminated entries at each round}
  }
}
\description{
Instant runoff voting.
}
\examples{
votes <- list(
  dex = c("Ice Skating", "Unihoc", "Food"),
  dean = c("Ice Skating", "Unihoc", "Food"),
  paul = c("Whiskey Tasting", "Established"),
  james = c("Ice Skating", "Unihoc", "Food")
)

irv(votes)

map <- c("e", "f", "i", "u", "w")
votes <- list(
  ballot(0, 3, 1, 2, 0, map = map),
  ballot(0, 3, 1, 2, 0, map = map),
  ballot(2, 0, 0, 0, 1, map = map),
  ballot(0, 3, 1, 2, 0, map = map)
)
irv(votes)
}
