% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series_daily_adjusted}
\alias{time_series_daily_adjusted}
\title{returns market end of day adjusted data for a specified symbol}
\usage{
time_series_daily_adjusted(symbol, outputsize = "compact", from = NULL,
  to = NULL, lastN = 100)
}
\arguments{
\item{symbol}{string, a market data symbol such as "AMZN"}

\item{outputsize}{depreciated}

\item{from}{date YYYY-MM-DD Optional return data starting at this date}

\item{to}{date YYYY-MM-DD Optional return data upto and including this date}

\item{lastN}{100 Optional; if 0 all data returned; if >0 lastN specifies how many of the most recent trading days will be returned}
}
\value{
dataframe
}
\description{
returns market end of day adjusted data for a specified symbol
}
