% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series_intraday}
\alias{time_series_intraday}
\title{returns intraday market data for a specified symbol}
\usage{
time_series_intraday(symbol, outputsize = "full", interval = "1min",
  from = NULL, to = NULL, nDays = 0)
}
\arguments{
\item{symbol}{string, a market data symbol such as "AMZN"}

\item{outputsize}{Depreciated}

\item{interval}{"1min"; choices are  1min, 5min, 15min, 30min, 60min}

\item{from}{date YYYY-MM-DD Optional return data starting at this date}

\item{to}{date YYYY-MM-DD Optional return data upto and including this date}

\item{nDays}{Optional 0 == all data; if >0 only most recent nDays of intraday data will be returned}
}
\value{
dataframe
}
\description{
returns intraday market data for a specified symbol
}
